fix: add tests for sync test results
