+++ tasks/tests/unit/conftest.py
@@ -1,11 +1,18 @@
 # TODO: Clean this
 
+
 import pytest
+from shared.django_apps.core.tests.factories import (
+    CommitFactory,
+    PullFactory,
+    RepositoryFactory,
+)
+from shared.django_apps.reports.models import Test, TestInstance
+from shared.django_apps.reports.tests.factories import UploadFactory
 from shared.reports.readonly import ReadOnlyReport
 from shared.reports.resources import Report, ReportFile, ReportLine
 from shared.utils.sessions import Session
 
-from database.tests.factories import CommitFactory, PullFactory, RepositoryFactory
 from services.comparison import ComparisonProxy
 from services.comparison.types import Comparison, FullCommit
 from services.repository import EnrichedPull
@@ -154,3 +161,67 @@ def sample_comparison(dbsession, request, sample_report):
             ),
         )
     )
+
+
+@pytest.fixture
+def repo_fixture():
+    return RepositoryFactory()
+
+
+@pytest.fixture
+def upload_fixture():
+    return UploadFactory()
+
+
+@pytest.fixture
+def create_upload_func():
+    def create_upload():
+        return UploadFactory()
+
+    return create_upload
+
+
+@pytest.fixture
+def create_test_func(repo_fixture):
+    test_i = 0
+
+    def create_test():
+        nonlocal test_i
+        test_id = f"test_{test_i}"
+        test = Test(
+            id=test_id,
+            repository=repo_fixture,
+            testsuite="testsuite",
+            name=f"test_{test_i}",
+            flags_hash="",
+        )
+        test.save()
+        test_i = test_i + 1
+
+        return test
+
+    return create_test
+
+
+@pytest.fixture
+def create_test_instance_func(repo_fixture, upload_fixture):
+    def create_test_instance(
+        test, outcome, commitid=None, branch=None, repoid=None, upload=upload_fixture
+    ):
+        ti = TestInstance(
+            test=test,
+            repoid=repo_fixture.repoid,
+            outcome=outcome,
+            upload=upload,
+            duration_seconds=0,
+        )
+        if branch:
+            ti.branch = branch
+        if commitid:
+            ti.commitid = commitid
+        if repoid:
+            ti.repoid = repoid
+        ti.save()
+        return ti
+
+    return create_test_instance
